Make immutable board model class.
