+++ neuralknight/scripts/cli.py
@@ -1,7 +1,7 @@
 from cmd import Cmd
 import requests
 
-from ..models import Board
+from ..models.board_model import BoardModel
 
 PORT = 8080
 API_URL = 'http://localhost:{}'.format(PORT)
@@ -32,7 +32,7 @@ class CLIAgent(Cmd):
         """
         Init player board.
         """
-        self.board = Board()
+        self.board = BoardModel()
         self.piece = None
         self.user = None
         game = requests.post(API_URL + '/v1.0/games').json()
