+++ neuralknight/tests/test_view_nofound.py
@@ -1,5 +1,5 @@
 def test_default_behavior_of_home_view(dummy_request):
-    from neuralknight.views.home import home_view
+    from ..views.home import home_view
 
     response = home_view(dummy_request)
     assert dummy_request.response.status_code == 200
@@ -8,7 +8,7 @@ def test_default_behavior_of_home_view(dummy_request):
 
 
 def test_not_found_behavior(dummy_request):
-    from neuralknight.views.notfound import notfound_view
+    from ..views.notfound import notfound_view
 
     response = notfound_view(dummy_request)
     assert dummy_request.response.status_code == 404
